+++ echo/tcp-echo/client/app/handler.go
@@ -32,30 +32,29 @@ type clientEchoSession struct {
 	reqNum  int32
 }
 
-type EchoMessageHandler struct {
-}
+type EchoMessageHandler struct{}
 
 func newEchoMessageHandler() *EchoMessageHandler {
 	return &EchoMessageHandler{}
 }
 
-func (this *EchoMessageHandler) OnOpen(session getty.Session) error {
+func (h *EchoMessageHandler) OnOpen(session getty.Session) error {
 	client.addSession(session)
 
 	return nil
 }
 
-func (this *EchoMessageHandler) OnError(session getty.Session, err error) {
+func (h *EchoMessageHandler) OnError(session getty.Session, err error) {
 	log.Info("session{%s} got error{%v}, will be closed.", session.Stat(), err)
 	client.removeSession(session)
 }
 
-func (this *EchoMessageHandler) OnClose(session getty.Session) {
+func (h *EchoMessageHandler) OnClose(session getty.Session) {
 	log.Info("session{%s} is closing......", session.Stat())
 	client.removeSession(session)
 }
 
-func (this *EchoMessageHandler) OnMessage(session getty.Session, pkg interface{}) {
+func (h *EchoMessageHandler) OnMessage(session getty.Session, pkg interface{}) {
 	p, ok := pkg.(*EchoPackage)
 	if !ok {
 		log.Error("illegal packge{%#v}", pkg)
@@ -66,7 +65,7 @@ func (this *EchoMessageHandler) OnMessage(session getty.Session, pkg interface{}
 	client.updateSession(session)
 }
 
-func (this *EchoMessageHandler) OnCron(session getty.Session) {
+func (h *EchoMessageHandler) OnCron(session getty.Session) {
 	clientEchoSession, err := client.getClientEchoSession(session)
 	if err != nil {
 		log.Error("client.getClientSession(session{%s}) = error{%#v}", session.Stat(), err)