+++ echo/udp-echo/server/app/readwriter.go
@@ -28,7 +28,7 @@ func NewEchoPackageHandler() *EchoPackageHandler {
 	return &EchoPackageHandler{}
 }
 
-func (this *EchoPackageHandler) Read(ss getty.Session, data []byte) (interface{}, int, error) {
+func (h *EchoPackageHandler) Read(ss getty.Session, data []byte) (interface{}, int, error) {
 	var (
 		err error
 		len int
@@ -49,7 +49,7 @@ func (this *EchoPackageHandler) Read(ss getty.Session, data []byte) (interface{}
 	return &pkg, len, nil
 }
 
-func (this *EchoPackageHandler) Write(ss getty.Session, udpCtx interface{}) error {
+func (h *EchoPackageHandler) Write(ss getty.Session, udpCtx interface{}) error {
 	var (
 		ok        bool
 		err       error