+++ echo/ws-echo/change_log.md
@@ -10,6 +10,10 @@
 ## develop history ##
 ---
 
+- 2018/06/24
+    > improvement
+	* delete this, using name abbreviation instead.
+
 - 2018/03/17
     > improvement
     * use getty 0.8.2